refactor: move some ballot metadata types to `types-rs`
