+++ libs/ballot-interpreter/Cargo.toml
@@ -38,10 +38,8 @@ serde_json = { workspace = true }
 serde_with = { workspace = true }
 itertools = { workspace = true }
 rqrr = { workspace = true }
-hex = { workspace = true }
 zbar-rust = { workspace = true }
 base64 = { workspace = true }
-static_assertions = { workspace = true }
 thiserror = { workspace = true }
 types-rs = { workspace = true }
