+++ libs/ballot-interpreter/src/hmpb-rust/interpret.rs
@@ -15,16 +15,17 @@ use rayon::iter::IntoParallelRefIterator;
 use rayon::iter::ParallelIterator;
 use serde::Serialize;
 use serde_with::DeserializeFromStr;
+use types_rs::ballot_card::BallotSide;
 use types_rs::election::{BallotStyleId, Election, MetadataEncoding, PrecinctId};
 use types_rs::geometry::PixelPosition;
 use types_rs::geometry::{PixelUnit, Size};
+use types_rs::hmpb;
 
 use crate::ballot_card::get_matching_paper_info_for_image_size;
 use crate::ballot_card::load_ballot_scan_bubble_image;
 use crate::ballot_card::BallotCard;
 use crate::ballot_card::BallotImage;
 use crate::ballot_card::BallotPage;
-use crate::ballot_card::BallotSide;
 use crate::ballot_card::Geometry;
 use crate::ballot_card::PaperInfo;
 use crate::debug::ImageDebugWriter;
@@ -33,7 +34,6 @@ use crate::image_utils::find_scanned_document_inset;
 use crate::image_utils::Inset;
 use crate::layout::build_interpreted_page_layout;
 use crate::layout::InterpretedContestLayout;
-use crate::metadata::hmpb;
 use crate::qr_code;
 use crate::scoring::score_bubble_marks_from_grid_layout;
 use crate::scoring::score_write_in_areas;
