--- libs/ballot-interpreter/src/hmpb-rust/metadata/mod.rs
@@ -1,3 +1,0 @@
-pub mod coding;
-pub mod hmpb;
-pub mod types;
